Now computes length and GC-content w/o considering introns.
However, the biomart and org.db methods don't behave in the same way.
In the biomart mode the length and gc-content of the longest isoform
is reported. In the org.db mode the length and gc-content of the
union of all isoforms is reported.

This need to be fixed, but it is not clear which is the best way.